BERTResearch/requirements.txt
datasets
praw
psaw
pandas
python-dotenv
tqdm
sentence-transformers 
scikit-learn
torch
matplotlib
seaborn
accelerate
tiktoken 
protobuf
sentencepiece